Move administrator login to the current page object pattern
