+++ tests/_support/Step/Acceptance/Administrator/Login.php
@@ -1,0 +1,30 @@
+<?php
+namespace Step\Acceptance\Administrator;
+
+use Page\Acceptance\Administrator\LoginPage;
+use Page\Acceptance\Administrator\ControlPanelPage;
+
+class Login extends \AcceptanceTester
+{
+	/**
+	 * @When I Login into Joomla administrator with username :arg1 and password :arg1
+	 */
+	public function loginIntoJoomlaAdministrator($username, $password)
+	{
+		$I = $this;
+		$I->amOnPage('administrator/');
+		$I->fillField(LoginPage::$usernameField, $username);
+		$I->fillField(LoginPage::$passwordField, $password);
+		$I->click(LoginPage::$loginButton);
+	}
+
+	/**
+	 * @Then I should see the administrator dashboard
+	 */
+	public function iShouldSeeTheAdministratorDashboard()
+	{
+		$I = $this;
+		$I->waitForPageTitle(ControlPanelPage::$pageTitle, 60, ControlPanelPage::$pageTitleContext);
+		$I->see(ControlPanelPage::$pageTitle, ControlPanelPage::$pageTitleContext);
+	}
+}
