+++ tests/acceptance/administratorlogin.feature
@@ -5,5 +5,5 @@ Feature: administrator login
 
   Scenario: Login in Administrator
     Given Joomla CMS is installed
-    When Login into Joomla administrator with username "admin" and password "admin"
-    Then I see administrator dashboard
+    When I Login into Joomla administrator with username "admin" and password "admin"
+    Then I should see the administrator dashboard
