+++ tests/acceptance/content.feature
@@ -5,8 +5,8 @@ Feature: content
 
   Background:
     Given Joomla CMS is installed
-    When Login into Joomla administrator with username "admin" and password "admin"
-    Then I see administrator dashboard
+    When I Login into Joomla administrator with username "admin" and password "admin"
+    Then I should see the administrator dashboard
 
   Scenario: Create an Article
     Given There is a add content link
